requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
httpx
python-dotenv
pydantic>=2.0
pydantic-settings>=2.0
semantic-kernel>=1.30.0
click
starlette
sse-starlette
jwcrypto
debugpy
